+++ src/main/java/com/skyteam/shopping_area/service/impl/UserServiceImpl.java
@@ -83,7 +83,7 @@ public class UserServiceImpl implements UserService {
         if (userRepository.findUserByUsername(user.getUsername()).get().getImage() == null) {
             newImage = imageService.saveImage(file);
         } else {
-            newImage = imageService.updateImage(file, Long.parseLong(user.getImage()));
+            newImage = imageService.updateImage(file, Integer.parseInt(user.getImage()));
         }
         user.setImage(newImage.getId());
         userRepository.save(user);
